requirements.txt
PyGithub==2.3.0
httpx==0.27.2
